move security controller back to the old location
